Switch url rewriter from js to typescipt
I pulled in a new library @aws-cdk/aws-lambda-nodejs that handles the
transformation of javascript to typescript.  Had to patch up a couple of
things which made this change a little more complex that intended.

- All of the CDK components were not version locked, that meant
  depending on the order of npm installs they could drift and get into a
dependancy mismatch spiral.  The version itself isn't important, its
uniformity is all that matters.

- Added permissions to the lambda role for interacting with cloudfronts
  distribution system, these were handled by the lambda function during
deployment, but it seems this new library while handling the js -> ts
brought some limitations and I had to manually implement the permssions see
for details https://tinyurl.com/5xy3n8zt

- Need to fix some typescript compliation issues, please feel free to
  re-work all of this and just keep the method signature

Signed-off-by: Peter Nied <petern@amazon.com>